# docs/OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Healthcare Services Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/doctors/register": {
      "post": {
        "tags": [
          "Doctor Onboarding & Practice Management"
        ],
        "summary": "Register Doctor",
        "description": "Submits doctor professional profile attached to Phase 01 User identity.\nInitial status is 'submitted'.",
        "operationId": "register_doctor_api_v1_doctors_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/documents/presign": {
      "post": {
        "tags": [
          "Doctor Onboarding & Practice Management"
        ],
        "summary": "Presign Document Upload",
        "description": "Generates pre-signed S3 upload URL for medical registration certificates or degrees.",
        "operationId": "presign_document_upload_api_v1_doctors_documents_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPresignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPresignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/documents/confirm": {
      "post": {
        "tags": [
          "Doctor Onboarding & Practice Management"
        ],
        "summary": "Confirm Document Upload",
        "description": "Registers the uploaded document metadata in the database against the doctor profile.",
        "operationId": "confirm_document_upload_api_v1_doctors_documents_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/me": {
      "get": {
        "tags": [
          "Doctor Onboarding & Practice Management"
        ],
        "summary": "Get Doctor Profile",
        "operationId": "get_doctor_profile_api_v1_doctors_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/me/documents": {
      "get": {
        "tags": [
          "Doctor Onboarding & Practice Management"
        ],
        "summary": "List Doctor Documents",
        "operationId": "list_doctor_documents_api_v1_doctors_me_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoctorDocumentRead"
                  },
                  "type": "array",
                  "title": "Response List Doctor Documents Api V1 Doctors Me Documents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/doctors/me/clinic": {
      "post": {
        "tags": [
          "Doctor Public Profile & Practice"
        ],
        "summary": "Save Clinic",
        "operationId": "save_clinic_api_v1_doctors_me_clinic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicCreateOrUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/me/toggle-listing": {
      "post": {
        "tags": [
          "Doctor Public Profile & Practice"
        ],
        "summary": "Toggle Listing",
        "operationId": "toggle_listing_api_v1_doctors_me_toggle_listing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/me/toggle-video": {
      "post": {
        "tags": [
          "Doctor Public Profile & Practice"
        ],
        "summary": "Toggle Video",
        "operationId": "toggle_video_api_v1_doctors_me_toggle_video_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/doctors/{doctor_id}": {
      "get": {
        "tags": [
          "Doctor Public Profile & Practice"
        ],
        "summary": "Public Profile",
        "operationId": "public_profile_api_v1_doctors__doctor_id__get",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doctor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorPublicProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reviews/doctor/{doctor_id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Doctor Reviews",
        "operationId": "doctor_reviews_api_v1_reviews_doctor__doctor_id__get",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Doctor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewRead"
                  },
                  "title": "Response Doctor Reviews Api V1 Reviews Doctor  Doctor Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClinicCreateOrUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "name",
          "address",
          "city",
          "locality",
          "pincode"
        ],
        "title": "ClinicCreateOrUpdate"
      },
      "ClinicRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "google_maps_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Maps Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "address",
          "city",
          "locality",
          "pincode",
          "id",
          "doctor_id"
        ],
        "title": "ClinicRead"
      },
...
      "DoctorDocumentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctor_id",
          "doc_type",
          "file_name",
          "s3_key",
          "uploaded_at"
        ],
        "title": "DoctorDocumentRead"
      },
      "DoctorPublicProfile": {
        "properties": {
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "years_experience": {
            "type": "integer",
            "title": "Years Experience"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "in_person_fee": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "In Person Fee"
          },
          "video_fee": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Video Fee",
            "default": "0.00"
          },
          "listing_online": {
            "type": "boolean",
            "title": "Listing Online"
          },
          "video_enabled": {
            "type": "boolean",
            "title": "Video Enabled",
            "default": false
          },
          "rating": {
            "type": "number",
            "title": "Rating",
            "default": 4.8
          },
          "qualifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Qualifications"
          },
          "clinic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClinicRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "google_maps_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Maps Url"
          }
        },
        "type": "object",
        "required": [
          "doctor_id",
          "full_name",
          "specialty",
          "years_experience",
          "in_person_fee",
          "listing_online"
        ],
        "title": "DoctorPublicProfile"
      },
      "DoctorRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "medical_reg_number": {
            "type": "string",
            "title": "Medical Reg Number"
          },
          "council_name": {
            "type": "string",
            "title": "Council Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "years_experience": {
            "type": "integer",
            "title": "Years Experience"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "listing_online": {
            "type": "boolean",
            "title": "Listing Online"
          },
          "video_enabled": {
            "type": "boolean",
            "title": "Video Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "full_name",
          "medical_reg_number",
          "council_name",
          "specialty",
          "years_experience",
          "bio",
          "verification_status",
          "listing_online",
          "video_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "DoctorRead"
      },
      "DoctorRegisterRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Full Name"
          },
          "medical_reg_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Medical Reg Number"
          },
          "council_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Council Name"
          },
          "specialty": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Specialty"
          },
          "years_experience": {
            "type": "integer",
            "maximum": 70.0,
            "minimum": 0.0,
            "title": "Years Experience",
            "default": 0
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "medical_reg_number",
          "council_name",
          "specialty"
        ],
        "title": "DoctorRegisterRequest"
      },
...
      "DocumentConfirmRequest": {
        "properties": {
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          }
        },
        "type": "object",
        "required": [
          "doc_type",
          "file_name",
          "s3_key"
        ],
        "title": "DocumentConfirmRequest"
      },
      "DocumentPresignRequest": {
        "properties": {
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "File Name"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "application/pdf"
          }
        },
        "type": "object",
        "required": [
          "doc_type",
          "file_name"
        ],
        "title": "DocumentPresignRequest"
      },
      "DocumentPresignResponse": {
        "properties": {
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "upload_url",
          "s3_key"
        ],
        "title": "DocumentPresignResponse"
      },
      "DocumentType": {
        "type": "string",
        "enum": [
          "medical_reg_cert",
          "degree_cert",
          "clinic_reg",
          "identity_proof"
        ],
        "title": "DocumentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "appointment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Appointment Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "doctor_id",
          "patient_id",
          "rating",
          "review_text",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewRead"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "published",
          "flagged",
          "hidden",
          "removed"
        ],
        "title": "ReviewStatus"
      },
...
      "ToggleListingRequest": {
        "properties": {
          "listing_online": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Online"
          }
        },
        "type": "object",
        "title": "ToggleListingRequest"
      },
      "ToggleVideoRequest": {
        "properties": {
          "video_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Enabled"
          }
        },
        "type": "object",
        "title": "ToggleVideoRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "under_review",
          "info_requested",
          "verified",
          "rejected",
          "suspended"
        ],
        "title": "VerificationStatus",
        "description": "PRD Section 2D: 6-state doctor verification pipeline"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}